Fixed bug in trigger code
The trigger registry caches the dependency graph computation
to figure out which triggers are relevant for a particular
table.  Previously we failed to tell the trigger registry
to update that cache when the rules changed.  This caused
us to miss triggers that were relevant.

This change fixes that bug.  It includes several additional
tests.  It also fixes a corner case
where inserting and updating didn't produce the same
tablenames, hence leading to odd behavior for trigger
tests.  This fix revealed a bug in the tests for
builtins, whose fix is also included.

Change-Id: I548951e155cb69d52920d28615d16916af1124ed