Disallow deletion of datasource if dangling references
Previously, you could delete a datasource even if there
were policy statements referencing the tables in that
datasource.

This change ensures it is impossible to delete a datasource
if you still have references to that table provided by that
datasource.  This change fixes a bug where Congress
fails to start because it does not have a schema for the
rules contained in the database, which can only happen
if the user has deleted the datasource with rules that
still reference it.

Closes-bug: 1435592
Change-Id: I040f2829fc284b1ddb954b757cea886a498171b4