Add skeleton for VM-placement engine driver
This change adds a policy engine driver base class
and a skeleton for a VM-placement engine.

Instead of having N+1 policies to represent tables
from N datasources and 1 policy for user-defined tables,
the VM-placement engine driver uses 1 policy to store
everything.  This was done for simplicity but required
a few changes in our datalog implementation.

Change-Id: I1ea16cf28e72b3c57fea31ee6847f8983095326c