Fix for graph dependency bug
There were 2 separate, but closely related bugs that caused
an uncaught exception when checking for cycles through rules
via the dependency graph computation.

Bug 1: initialize_tables() doesn't change the dependency graph,
       but update() (i.e. insert/delete/update) does.
Bug 2: modifying the dependency graph even when the update
       is a noop

These bugs made it possible to delete a node in the (refcounted
version of the) dependency graph, even though there was still an
edge pointing to that node.  Running DFS on such a graph causes
an uncaught exception.

This change eliminates the bugs and adds tests to catch such
problems in the future.

Closes-bug: 1435842
Change-Id: Ibb0034338d3d2464f0ecf34935c706f51ac44110