Fix inconsistent rule insert bug
Previously, if you inserted 2 facts into the policy engine with
different numbers of columns, you would get an error that would
leave the system in an inconsistent state where the 2nd fact
is in the policy engine, but is not indexed, and is not in
the database.

This fix catches exceptions thrown during indexing and reverts
the change to the policy-engine before re-raising the exception.
This fix also includes a test for this flaw.

Related-bug: 1438398
Change-Id: Id2abf06da68cf87cee7a8754564e143fe10db807