Add Datalog to LP conversion
This change converts Datalog into a PuLP LP program.
The conversion consists of 2 steps: (i) converting Datalog to
an arithmetic constraint language that understands and, or, =, !=,
arithmetic, and variables and (ii) converting that language
into a linear program.

For the use cases we target, the arithmetic constraint
language is semantically the same as a linear program, except that
the constraint language is allowed to include and/or
and boolean indicator variables (like Y = X > 2).
In contrast, in a linear program every formula is of
the form:

ax1 + ... + axn [<=, =, >=] b.

Change-Id: Ic298b3b2c9e655cac23b65642a450a3dcc6a3312