Fix bug in action theory on restart
The bug was that creating a theory of kind 'action' produced
an in-memory 'action' policy but stored the type of that policy
as 'nonrecursive' in the database.  So on restart, the policy
engine would create a policy of the wrong kind, which would
in turn cause some rules stored in the DB to be incorrectly
rejected, thereby causing a fatal error.

The fix was to change the code that looked up the type of
a theory from checking isinstance() to looking at the
field of the theory that is set to the name of the policy
type.

Change-Id: I869fa835d07a863e8d5473a638db52e3ed4d6133
Closes-bug: 1439536