Add murano environments to murano datasource tables and fix bugs
Includes environment in the parent_types and relationships
datasource tables. Also fixes a bug where parent_types entries of
an application/service are missing if its type prefix is not
'io.murano.apps' as in the case of MySql ('io.murano.databases.mysql')
Now service type is checked against loaded application packages so
that the service type is properly identified. If its type is not
found then its type prefix is used for identifying its type as the
last resort.

Implements: blueprint murano-driver

Change-Id: I9b5fcbade1398678238ac6b470d4eccdd1ffc893