Add parser/lexer
These files are generated from congress/datalog/Congress.g,
but (i) they change rarely and so no one remembers to run
make and (ii) they require Java to generate.  So it's
better for end-users and most developers to check the files in
and just have the people changing the grammar remember to
check in the generated files.

Closes-bug: 1387209
Change-Id: Ia315fc28570edfaa7d8b223c46f514fb7e34921e