Improve arity computation
The arity computation was overly complex, and it sometimes failed to
compute the answer when modals were involved.  This change makes
the arity computation simpler and more robust.

Part of that simplicity comes from eliminating
the ability to compute arity using theories included
in each other.  But it seems we are not using that capability
since theory inclusion is only used by simulation,
and the simulation tests pass.

In the future, we should try to completely eliminate theory
inclusion if possible.

Change-Id: Ic376f92515621e29297ca456e1cd424e18c374fc