Fix bug in execute[nova:p(x)]
If the name of the action was qualified by the name of the service
the contents of execute[nova:p(x)] was always empty and so the trigger
would fail to execute.

The root cause of the bug was in computing the name of this table for
the dependency graph: the name of this table was simply nova:p when
in reality it should have been <policy>:nova:p.

Change-Id: I7baa79569b5147dfe611af7b3e09a6a44a1593ee