Sort head and body before computing a Rule's hash
The policy engine may change the order of a Rule's body before inserting the
Rule into the policy engine.  The previous version computes the Rule's hash
based on the unsorted body, so if the body order changes before the Rule
computes the hash then a subsequent attempt to remove the same Rule will fail
because the hashes do not match.

This change sorts the head and body before computing the hash so that the hash
will stay the same even if the body order changes.

Also add a test that inserts and removes a rule, and checks that the rule is
no longer in the policy engine.

Change-Id: I0f46ee22dcf3a5facff279266e927fb1290a014e