Add test for execute[nova:action(...)]
Added an end-to-end test for writing 'execute' policy statements
and fixed the bugs that came up.

The test does the obvious thing:
user writes a policy statement with 'execute', and then the test
ensures that when appropriate the action is called on the appropriate
service.  To succeed, the policy engine must successfully send
a request over the DSE, and the service on the other side must
successfully execute the correct action. (Of course, the action
is mocked out, but that code must be run with the correct arguments.)

Change-Id: I93ab4fd5c26278a587165a70211bb3fe621fac31