Add validation to datasource creation
Previously, there was no validation performed when creating a datasource,
so if you failed to supply any --config options, you got an internal
server error.

This also fixes a bug where if the DSE creation-service fails, the code
fails to destroy the policy it already created before re-raising the
exception.

Change-Id: Ifd8e38ca5b9fead46e5d3fc9dcc4faa77ea42ab7