Remove support for policy snapshot file
The snapshot file support bypasses the database when inserting rules into the
policy engine, so it makes it possible to insert rules into the engine without
the rules being in the db.  That means if Congress tries to synchronize rules
with the db, the synchronizer will delete rules from the snapshot because the
rules do not exist in the db.

Change-Id: Icbd6d1e1f1233aa00157d036f697933da07347f7