Fix bug with policy negation and multiple policies
If negation is applied to an atom that references a different
policy, that negation always succeeds.  The root cause was
that negation caused the evaluation engine to skip over
the logic that runs the evaluation routine within another
policy.  This meant that the evaluation engine could
never prove anything positive in that other policy and
hence the negation always succeeded.

This change routes the logic through the normal evaluation
path.  This ensures that what is inside the negation is
evaluated exactly the same as if it were outside the
negation.

Change-Id: I2d5ceeccf87afd99adfe71c359cf2b5e5cd46b38
Closes-bug: 1448295