Add status api calls for policy and rule
This change allows an API client to request the status of a rule or policy.
The response will reflect the state of the policy engine, not the state of the
database, since the policy engine and the database can become out of sync.

For a policy that exists, the API response will contain the policy name and
id.  For a rule that exists, the API response will contain the rule name, id,
comment, and original string.  If the rule or policy does not exist, the
congress server will return a 404 error.

Change-Id: I5c134e14d1ae5fec802a649be9834f86dab5c37a