Fix bug when processing built-in semantics
If built-in semantics are used in datalog, the evaluation engine will
always fail because it skips the logic of built-in evaluation and report
None instead.

This change ensures that the result of built-in evaluation will report
correct answer in normal evaluation path.

Change-Id: I89a21147bfd617713f74778335a633e6f7166fa6
Closes-bug:1449374
(cherry picked from commit 4cb93270580eb76d8de2907ff0cd54618f92a8c5)