Fixes to Glance tempest tests
Improved error handling in the Python client caused
fatal errors in this test, instead of simply waiting
a bit longer. The error was due to the driver not
always having self.state set for all the relevant tables.

This fix ensures self.state is always set for all the
relevant tables.

Change-Id: I73a0c3b3c2bffa4d000712688740948d59171a3f