Set unique object ids in data source tables' table
The Horizon Column class caches objects by id. The ids of the data source
tables are not unique across data sources. When two data sources have a table
with the same id, this causes the cached value of the first object to be
used for the second object. The same value is then always shown in that column
for those two rows in the Data Sources page's DataTable. Fix that by setting a
unique id for each data source table object and storing the actual table id
under another key. Similarly for policy tables on that page.

Change-Id: I6f8a7f05ed9f22013cf9690e89ddac1c1624d920
Closes-Bug: #1413726