Add id function extractor to datasource driver
Previously, there was no way to build a custom identifier
when translating an API call result into tables.  The
only options were to specify a column name or to have
an ID generated for you.

This change makes it possible to specify a function that
given the object returns an ID.

Implements-blueprint: datasource-driver-custom-id
Change-Id: I9ff134ef56fb118ba846fcdef11a656e25435df6