remove deprecated callback in schedule()
In deepsix, when a subscribe or request attempt needs a callback, the
callback is registered in `self.subdata[msg.correlationId]`.
So when a 'pub' or 'rep' message is received later, the callback can be
get from `self.subdata[msg.correlationId].callback`, then be called.
It is unnecessary to pass the callback to `schedule()` method, and this
argument is never used in `schedule()` either.

Change-Id: Ib510025487ca93cd14fac04ee5717628b4bcc1a0
Closes-Bug: #1475473