Prepare Policy Model API for RPC
Previously the policy model API used Python pointers to the
policy engine for its implementation.  This won't work when
the API model is running in a different process than the
policy engine.  Also, the API model was responsible for
keeping the in-memory policy engine in sync with the database;
this is unnatural once the policy engine is running in
its own process and using its own database.

This change rewrites the policy model so that it uses
a fake RPC method instead of a direct Python reference to the
policy engine.  It also moves the database synchronization to
the policy engine.

Partially-implements: blueprint dist-api-rpcify-policy
Change-Id: I87116cc43bd280788bb810db6d668f0ca7f4b07a