Add an object extract function key to translators
This allows datasource translators to preprocess any type of data
input for DataSourceDriver.convert_objs(). Translator newly has a
key, named 'objects-extract-fn',  which value has python reference
to a converting method. The method is defined by datasource
developers and converts an input of convert_objs() to list, dict,
or individual value.

For example, if a datasource driver retrieves a string of JSON
and it's passed to convert_objs, the preprocess method change the
string to Python dict.

Change-Id: I05691b06e95ef300bfe5c94d334b54e076a373c0
Implements: blueprint add-extract-fn-to-all-translators