RPCify rule_model
Previously the rule_model manipulated the methods of the
policy engine directly and maintained consistency between
the database and policy engine.  Once the API and policy
engine are running in separate processes, that will no
longer work.

This change introduces the RPC abstraction for the rule_model.
It also moves the maintanence of the policy engine and
database into the policy engine itself, effectively
creating a persistance later for the policy engine.

Also moved a few functions around in the policy engine
to make the layering clearer.

Partially-implements: blueprint dist-api-rpcify-rule
Change-Id: I2be8ef35dd9819036a24e617b3ad1aa268aa2782