Add native column-references into policy engine
Previously the policy engine assumed it had the schemas for
all datasources before it loaded rules from the database
or processed rule insertions from the user.  In particular,
any rule with a column-reference was compiled into a rule
without column-references, an operation that required
having the schemas available.

In the distributed architecture, we can no longer assume
the policy engine has access to schemas before rules
are loaded.  Schemas are sent via RPC calls from
datasources when they spin up.  Without any changes
to the policy engine to deal with schemas, the policy
engine would fail to boot if there were any rules
with column references.

This change makes the policy engine flexible enough to
load and process rules before the schemas arrive.  In
particular, the policy engine will now disable any
rule containing a column-reference for which the
schema is unknown.  Once all the appropriate schema
arrive, disabled rules are enabled.

Limitations of this change
- Schemas are set-once, i.e. once they are set they cannot
be set again.
- The policy engine does not check rules against the schemas
when they are set, except for disabled rules.
- Once a rule is disabled, the policy engine should probably
disable other rules because they rely on tables that are not
fully defined, but this functionality is not yet present.
- It is possible for a schema to come in that is inconsistent
with an existing disabled rule.  Such rules are stored in an
'error' datastructure, but there is no channel to the user to
inform them of these errors because the set-schema event
is caused not by the user but by the datasource driver.

Change-Id: I6aa1ff24d675d553d80a7d0fb9177fa8169e27a4