Unbreak master with 2 fixes in 1 commit
1. Avoid double-use of 'name' value in exceptions

Previously the Congress Exception used 'name' as the name of
the error code to use and some of its subclasses used 'name'
as a placeholder in the error message.  This caused problems
because CongressException popped 'name' off the keyword args
in the super class, which meant it was unavailable during
message construction.  This was causing congress to fail
to start in devstack.

This change eliminates the use of 'name' in the subclasses.

2. Fix nova datasource driver instantiation failure

The nova client failed to load because an equality check
on nova client was raising an exception.  This was causing
all unit tests using nova to fail.

This fix replaces an implicit == check with an 'is' check.

Closes-bug: 1492334
Closes-bug: 1492051
Change-Id: I8c166cd4db6d4c45d0709cb33a16bc3188d3d5eb