Ignore KeyError for secret field
Although it doesn't make sense to create a datasource without
providing auth info in config info, it may be not possible to
define a schema to verify because different datasource drivers
may use different kinds of auth methods. So we should be able
to handle the case that none secret info is provided. The fix
just simply ignores KeyError exception here.

Change-Id: I06bc6a4fb659d862da96dd96557454b6385d003f
Closes-Bug: #1492354