Ensure datasources are properly deleted
Previously, deleting a datasource driver did not remove it from the DSE.
That was caused by the polling thread having a reference to the object
it was polling and therefore the object was not properly garbage
collected upon deletion.

This change deletes the polling thread and then deletes the object,
which allows garbage collection to properly eliminate the datasource.

Change-Id: I8063dab0786e7c728ba0c0c9b3cc33748f70a38c
Closes-bug: 1493430