Replace local copy of dateutil 1.5, use dependency
Having a local copy of a library is a bad practice. Files are
duplicated on the system, it's harder to update it.

For example, the local copy of dateutil 1.5 is not compatible with
Python 3.

Replace the local copy with a dependency to dateutil, at least the
version 2.4.2. dateutil 2.4.2 is compatible with Python 3.

Partially implements blueprint: support-python3
Change-Id: I366c8004e5eb683b41fec2c31cf4a5afdb48180c