Fix state computation incorrect
The Resource of oslo-incubator is used as the
super class of client object in novaclient and
cinderclient, but the implemented of __eq__ method
don't match the semantics of object equal. Check
equal between two objects by using obj1 == obj2,
that will always be True if the objects have the
same id, even if they have the different other
attribute value. The issue had been traced by
oslo-incubator bug/1499369, fix it in Congress side
as temporary workaround in order to avoid the issue
exist in stable/liberty branch.

Change-Id: I4304c32432dc1f813377c60704f817b4f3020da2
Closes-Bug: #1499065