Change devstack to OS_PROJECT_NAME and avoid mocking oslo.policy internals
To pass the tests in gate, this commit has 2 separate changes

1) OS_TENANT_NAME is no longer used in the latest devstack, it causes all
datasource driver fail, as well as jenkins gate failure.

This fix changes devstack to use OS_PROJECT_NAME

2) Previously, we tested oslo.policy by mocking out one of the internals
of oslo.policy, just like Neutron.  Besides being improper, the newest
release of oslo.policy broke that test, which is what is blocking
the tip of master currently.

This fix shamelessly copies the fix out of Neutron, mocking out
at the layer of HTTPCheck.

Change-Id: I2a4eae714946ce98a060a43dd501b9b3619aa68e
Closes-Bug: #1505057
Closes-bug: #1505750