clarify the example in 'Simulation with Actions'
In RST document congress/doc/source/enforcement.rst.

The policy 'aliceactions' is created at first in the example,
but that is never used, all the following command use the
default 'action' policy, that make the users confused about
which policy we should create "action('set')" rule into and
which action policy we should use in simulate? Aims to clarify
the usage, avoid to use default 'action' policy, and use
'aliceactions' policy in the example.

Change-Id: I1c976120c20b8e766a6692d3ab16e91f92da8b5c
Closes-Bug: #1510428