Add availability zone information into nova driver
The user can specify which AZ is used to boot up a vm. Since AZ can be
also used to separate security zones it is very useful to have this
information in Congress in order to monitor if VMs are placed in the
right network and uses the correct volume backend.

For more detail information see [1].

[1]: https://github.com/openstack/telcowg-usecases/blob/master/usecases/sec_segregation.rst

Change-Id: Ie53a0fa0600a20f478791a81e7c750378e91476b