Migrate API policy model to dse2
Made a copy of the policy-model API tests
and put them into tests2/api so that we can easily
load the new dse2 and modify the tests that don't
work verbatim on dse2.  Grep for TODO(dse2) for
comments on which tests haven't been moved over to
dse2.

Also had to migrate the rule-model because it was
used in theh policy-model tests.  But the rule-model
tests are not yet migrated over.

To run the new tests:
python3 -m "nose" congress/tests2/

Implements: blueprint dist-integrate-new-dse
Change-Id: Ic2eddc7ad3d03162d02be675546be1161a1b07c5