Add original tests that are passing to tests2
Some of the dse1 tests that are passing aren't run with the
distributed_arch flag set to True since they haven't been
copied to tests2.

This change adds stubs to tests2 that ensure they are being
run along with the tests2 tests.  The stubs are not copies
of the original tests--they just inherit from them.  So
the stubs are given different filenames to make this clear.

Implements: blueprint dist-integrate-new-dse
Change-Id: Ic49b3890ce2d924f0dcd09decf7eb97dc4ea5ea1