Raise proper exceptions in new architecture
This commit raises exception, if a call is invoked on invalid
service. This should enable the tests disabled as proper exception
is not raised for invalid datasource.

Added congress.exception class to allowed_remote_exmods, to reraise
the exceptions from remote endpoint with same type instead of
raising as RemoteError.

Change-Id: I1deb4d350e8be90c9c8a76e0b5c7c9e7051405a8
Closes-Bug:#1542540