Moved CORS middleware configuration into oslo-config-generator
The default values needed for congress's implementation of cors
middleware have been moved from paste.ini into the configuration
hooks provided by oslo.config. Furthermore, these values have been
added to the default initialization procedure. This ensures
that if a value remains unset in the configuration file, it will
fallback to using sane defaults. It also ensures that an operator
modifying the configuration will be presented with that same
set of defaults.

Change-Id: I2edc24e4d47c4d664dd31c407d46e42fefdb1488
Closes-Bug: #1551836