Moved handling of polling threads to PollingDataSourceDriver
After adding PushedDataSourceDriver, handling of polling threads was
left in the common base class (DataSourceDriver). Polling is not used
by PushedDataSourceDriver and it was causing problems when deleting
datasources (i.e. trying to kill thread which didn't exist).

With this change, polling thread handling code is moved down to
PollingDataSourceDriver class and logic is extended to check if the
thread has actually been created. Only then thread termination will
be attempted.

Change-Id: Ifa450fa6bd1c5e15937bbdc76acd9fad64f4d18e
Closes-Bug: #1553982