Reject rules with column refs on non-datasources
Previously, the policy engine would accept rules that had
column references on tables in policies that were not
datasource policies and disable them.  Unfortunately,
there was no way to enable those same rules because
there is no way to set the schema other than for
datasource policies.

This change throws an exception when someone tries
to insert a rule that has a column reference
for a table within a non-datasource policy.

Also, fixed bug where eliminate_column_references
for Literal and Rule were not fully polymorphic.
Literal and Rule expected different arguments, but
the callers were providing the same arguments.

Change-Id: Ia0f1a9980befb0189d5727dfcfacad376fab94fe
Closes-Bug:#1553433