raise exception on adding same rules with different literal order
currently, same rules with different literal order
are considered as diff rules if it has datasource tables. The
reason is while eliminating column references, we prepend the prefix
x_i for unnamed arguments, resulting in diff hash, hence diff rules.

This can be fixed either using the same prefix or by the sorting
the literals, which would be identified as same rules

This commit uses the same-prefix solution.

Change-Id: I4873de8be88894b28a4153a3d5bcb0cdefff9f9a
Closes-Bug:#1553430