Make persistent_insert_rule error on IncompleteSchemaException
Before this patch, persistent insert of rules with unknown table with
column-name references leads to unexpected behavior:
- misleading 'Rule already exists' error
- limbos the rule
- but never persists it in the DB, even if the rule is enabled by
  new schema.

With this patch, the rule is outright rejected with syntax error.
The rule is not limboed.

Additional notes:
This patch largely preserves the existing behavior on rule insertion.

But it may be helpful to review the the behavior:

When inserting a rule that refers to a table belonging to a policy,
we have the following four cases of interest:

1. policy is unknown (could be a datasource policy in the future)
2. policy is a known datasource policy, but the schema is not
   yet available
3. policy is a known datasource policy, and the schema is
   available, but table is not in the schema
4. policy is known not to be a datasource policy.

Pre-existing behavior:
For rule without column-name references, all four cases are accepted.
For rule with column-name references:
a. For non-persistent insert, (1-3) are limboed and (4) is rejected.
b. For persistent insert, (1-3) are limboed (with extraneous error)
   and (4) is rejected.

With the patch, all four cases are rejected in (b).

Neither the pre-existing behavior nor the patched behavior is
particularly consistent.
It may be appropriate to re-design the behavior in the future.

Terminology notes:
Accepted means rule added and put in effect.
Rejected means rule not added and error shown.
Limboed means rule is added to disabled events, to come into effect
under the right conditions (eg. matching schema received).

Closes-Bug: #1553433

Change-Id: I5b377cd7868004d7891263a0649a31621845668c