Fix Gate HA test Failure
This commit fixes the HA issue happenning from time to time.
The problem is with explict call to create/delete datasource
policy in add_datasource/delete_datasource which is causing
the error, if synchronizer is run between the 2 statements.

So, in this commit removed the explicit call to create/delete
policies and let synchronizer take care of the same in next
periodic iteration.

Haven't added explicit call to synchronizer because of following-
1. In case of HA, even if explicit call to synchronizer added
will only run on one PE node and other nodes needs to be
synchronized later (or we need to broadcast the call) ,
and also if the synchronizer periodic task is running in
different thread, and making an explicit call again, might
cause race condition to happen.

Change-Id: I76e8c859104eaa5d966c9c146dcf0f6bec372d71