Scrap testHA:test_datasource_db_sync_remove
The test is redundant because removal by db sync is already tested
in test_datasource_db_sync_add. The test also causes frequent
transient failures.

The following bug not completely closed because there appears to be
deeper issues causing the failure. To be addressed in coming days.

Submitting this quick fix to unjam the current gating process holding
up ready and to-be-ready patches.

Change-Id: If7013ef8afa9b3b9cbe773ba1393f5f822f60003
Partial-bug: 1611467