Add support to synchronize one policy
This commit adds supports to sync one policy only instead of
looping all the policies everytime. As this function is used
in both Runtime and DseRuntime, moved temporarily to parent class

TODO(ramineni): Will move the same into different class to be
able to use by both classes. leaving it for now, to avoid major
change.

Change-Id: I24968407c3daa3124814c11309076178c84bf6be