tempest HAHT wait for sync
Add a wait timer in tests that rely on replicated nodes to sync.
With this change, the gate-congress-pe-replicated-nv test is
expected to pass and this provides useful feedback.
Note: the existing retry logic do not work for some reason.
But we don't fix the retry logic because retry is undesirable for
this kind of testing (explained in comment).

Change-Id: Ic9adc1ddcb1e86e5ab75d1154b8bfa45b51f3fa1