Add namespace 'builtin' for builtins
Previously, builtins and helper tables were in the same namespace.
This is problematic because whenever we introduce a new builtin,
it would override any helper tables that a user had written
using that same name.

In this change, all builtins are prefixed with builtin: so we are free
to add as many new builtins as we like.  So that we can officially
deprecate the bare version of builtins (i.e. builtins not prefixed
by builtin:), the bare version is still allowed.  Once the deprecation
period is over, we will remove support for bare builtins.

Change-Id: I0438826289a4219b9d1043f8132ba5d193e8751b