Fix check that policy-names are valid IDs in the grammar
Previously, users could create a policy with a name that
could not be used as a servicename in the grammar.  There
was a check in place, but the test failed to catch all the
cases, e.g. where the servicename is a UUID.

This change fixes that check.
This change also adds a check for datasources to ensure
all datasource names can be used in the policy language.

Change-Id: I3a0bea72eae233d6e103b850512236b6938616a4