Add fields to cinder:volumes and new attachments table
The following information is added: encrypted, availability_zone,
replication_status, multiattach, snapshot_id, source_volid,
consistencygroup_id, migration_status, attachments.

We discussed on IRC how to deal with changes to the data schema. There
was some agreement that versioning the schema was the most complete
solution.

But at the moment and in the medium term, we do not have much need for
understanding the different versions in code (ie. a version parameter
gets passed to the driver when creating a datasource). So in this
patch, I propose to document the cinder driver schema version as 2.1
(and the original version 2.0 to coincide with the cinder client
version used.) Rules written for 2.0 is compatible with 2.1 driver.
But rules written for 2.1 schema may not be compatible with the 2.0
driver.

The schema version is helpful so that when a policy is documented
with the schema version it is based on, users can figure out whether
it is compatible with a particular install of congress drivers.

At the same time, we avoid some implementation and maintenance
complexity by maintaining only the latest minor version of each major
version.

Note: the backward compatibility of this change would be achieved with
positional argument padding under review here:
https://review.openstack.org/#/c/448828/

Closes-bug: 1674537

Change-Id: Ia0f2f17b4e25f70600175ababb35805fbcdd3142