Datasource for configuration file validation
This is the datasource itself that communicates with the agent through
the message bus.

Implementation of a datasource that transcribes the content
of configuration files managed by oslo-config in Congress
tables.

The datasource uses a set of agents deployed on the nodes
to access the configuration files.

Change-Id: I9919a3595225440cacdbfa9633c0cb5ccd7e31fd
Implements: blueprint configuration-files-validation
Co-Authored-By: Valentin Matton <vmatt.openstack@gmail.com>
Co-Authored-By: Pierre Crégut <pierre.cregut@orange.com>