Add attach/detach security group action to neutron driver
To modify the security groups attached to the list through congress
policy, actions to attach/detach a security group to/from port is
needed.

Because the neutron v2 API and client do not support attach/detach
of a single security group, but only the replacement of the entire
security group list, we add these custom actions which combine
a read and update action to accomplish the attach/detach.

Caveat: Another concurrent update to the list of attached ports may
be lost. Unfortunately, there is simply no better solution given the
limitations of the neutron v2 API. So we enable these actions anyway
with documented warning in the driver actions description.

Change-Id: I51ec7452a5a7a77cc8cfc57b87e671685280ea73