Use oslo.context from_dict unmodified
With the change to oslo.context in the depends-on, we no longer need
to maintain an override for the from_dict function.  We only need
to provide a list of additional keys in the appropriate class member.

This will work with oslo.context 2.19.2 and above, which is now the
minimum version in global-requirements.

Depends-On: Ib143f8a5c129dbf6711800c4d87c8830a8aa3365

Change-Id: If09a3b22465b196baa34edc00c7472f5feed4ba5