Fix serialization of choices and reenable congress-agent.
oslo_config uses OrderedDict for choices in
options types. This serialize it as list of tuples.
Order is  preserved and types are naturally
deserialized back from those tuples.

Congress-agent is reenabled. It can be disabled
in devstack on  the value of ENABLE_CONGRESS_AGENT
(default is True).

Driver will initialize table to empty at first poll not
waiting for an initial response from agents.

Closes-Bug: #1742833
Change-Id: If1423e824fc457bab89f4bcd12be8826c50f5f69