Data source driver preserves data types
Move the conversion to agnostic-compatible types from
datasource_driver to agnostic.

Internal change only; external API unaffected.

The datalog compiler and the agnostic policy engine in Congress
do not support booleans and Nones/nulls. To work with that design,
the data source drivers currently convert booleans and Nones into
strings. This works well for agnostic, but unnecessarily limits
the types handled by the data source drivers.

In the spirit of the original architecture to support policy engines
in addition to agnostic, this change expands the types handled by the
data source drivers and the Congress data bus to all the JSON scalar
types, de-coupling from the more limited set of types accepted by
agnostic. This way, other policy engines can independently choose
which subset of the data types to work with.

This would be a small first step toward integrating other policy
engines such as those based on SQL, z3, or Open Policy Agent.

Change-Id: Ia63614a84742833c75ca3cadec7db48a44200e9d