Expose a get_enforcer method for oslo.policy scripts
Because we have policy in code, we should be able to use the
oslo.policy CLI scripts to produce sample policy files and render
complete policies based on overrides on disk. This was broken
because Congress did not properly expose the appropriate end point.

This commit exposes a get_enforcer() that substitutes an empty list
in place of arguments passed in through the system.

Based on: https://review.openstack.org/#/c/530828/

Change-Id: Ia5f1f949b2dc97cce4227a8dd6819b4354867b97
Closes-Bug: 1793343