change the dist name to 'openstack-congress'
We are updating all Python projects to publish artifacts to PyPI. The name
"congress" is already taken there by another project, and they have not
responded to our request to claim the name. We therefore need to change
the dist name used to package congress. We have some other projects
publishing using an "openstack-" prefix, so I propose using the name
"openstack-congress". This will not change the imports or anything else
about how the code works, just how it is packaged.

Change-Id: I076a4025c40e71845e62fe1747cb7c33a477ea16
Signed-off-by: Doug Hellmann <doug@doughellmann.com>